edf_databroker_train/requirements.txt
grpcio==1.41.1
grpcio-tools==1.41.1
protobuf==3.19.0
six==1.16.0
